requirements.txt
argparse~=1.4.0
pyOpenSSL~=19.1.0
openpyxl~=3.0.4
requests~=2.24.0
cryptography~=2.9.2
python-dotenv~=0.14.0